Redesign tips panel
