+++ ui/client/components/tips/Warnings.js
@@ -1,0 +1,46 @@
+import InlinedSvgs from "../../assets/icons/InlinedSvgs"
+import React from "react"
+import HeaderIcon from "./HeaderIcon"
+import {v4 as uuid4} from "uuid";
+import {Link} from "react-router-dom"
+import NodeUtils from "../graph/NodeUtils"
+
+export default class Warnings extends React.Component {
+
+  render() {
+    const {warnings, showDetails, currentProcess} = this.props
+    const groupedByMessage = _.groupBy(warnings, warning => warning.error.message)
+    const separator = ', '
+
+    return (
+      <div key={uuid4()}>
+        {warnings.length > 0 && <HeaderIcon icon={InlinedSvgs.tipsWarning}/>}
+        <div>
+          {
+            Object.entries(groupedByMessage).map(([message, warnings]) =>
+              <div key={uuid4()}
+                   className={"warning-tips"}
+                   title={warnings.description}>
+                <span>{headerMessageByWarningMessage.get(message)}</span>
+                <div className={"warning-links"}>
+                  {
+                    warnings.map((warning, index) => <Link key={uuid4()}
+                                                           className={"node-warning-link"}
+                                                           to={""}
+                                                           onClick={event => showDetails(event, NodeUtils.getNodeById(warning.key, currentProcess))}>
+                      <span>{warning.key}</span>
+                      {(index < warnings.length - 1) ? separator : null}
+                    </Link>)
+                  }
+                </div>
+              </div>
+            )
+          }
+        </div>
+      </div>
+    )
+  }
+}
+
+const headerMessageByWarningMessage = new Map([["Node is disabled", "Node disabled:"]]);
+
