+++ ui/client/stylesheets/graph.styl
@@ -33,7 +33,7 @@ node-highlight(strokeColor, backgroundFill)
     fill: backgroundFill
 
 .node-validation-error
-  node-highlight(#f25c6e, #fbd2d6)
+  node-highlight(errorColor, #fbd2d6)
 
 .node-focused-with-validation-error
   node-highlight(#0058a9, #f2dede)
