+++ ui/server/Errors.md
@@ -50,4 +50,4 @@ Process warnings (not necessarily errors)
 
 | Error type                       | Message                              | Explanation                                                      |
 | :--------------------------------|:-------------------------------------| :----------------------------------------------------------------|
-| UIValidation                     | {$id} is disabled                     | Deploying process with disabled node can have unexpected consequences         |
+| UIValidation                     | Node is disabled                     | Deploying process with disabled node can have unexpected consequences         |
